# [PRW 2.0]: Migrate gogo/proto to vtproto