--- a/prompb/buf.yaml
+++ b/prompb/buf.yaml
@@ -17,5 +17,3 @@ lint:
       - remote.proto
       - types.proto
       - io/prometheus/client/metrics.proto
-deps:
-  - buf.build/gogo/protobuf