--- a/prompb/remote.proto
+++ b/prompb/remote.proto
@@ -17,14 +17,13 @@ package prometheus;
 option go_package = "prompb";
 
 import "types.proto";
-import "gogoproto/gogo.proto";
 
 message WriteRequest {
-  repeated prometheus.TimeSeries timeseries = 1 [(gogoproto.nullable) = false];
+  repeated prometheus.TimeSeries timeseries = 1;
   // Cortex uses this field to determine the source of the write request.
   // We reserve it to avoid any compatibility issues.
   reserved  2;
-  repeated prometheus.MetricMetadata metadata = 3 [(gogoproto.nullable) = false];
+  repeated prometheus.MetricMetadata metadata = 3;
 }
 
 // ReadRequest represents a remote read request.