--- a/scripts/tools.go
+++ b/scripts/tools.go
@@ -18,7 +18,6 @@
 package tools
 
 import (
-	_ "github.com/gogo/protobuf/protoc-gen-gogofast"
 	_ "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway"
 	_ "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger"
 	_ "golang.org/x/tools/cmd/goimports"