@cstyan 2024-09-16
Tony ( @tonyli233 ) is mentioning that some of the benchmark results are concerning/could still be improved, but it's worth noting that some bechmarks haven't been updated to use the new v2.0 format yet and others don['t work with v2 proto (such as BenchmarkSampleSend, which the v2 version never completes atm).

His benchmark results at the top of this PR are showing gogoproto v2 vs vtproto v2, but if we compare both to main with v1 we're still seeing good results. For example:
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: AMD Ryzen 9 5950X 16-Core Processor
              │    main     │               main-v2               │                vt20                 │
              │   sec/op    │   sec/op     vs base                │   sec/op     vs base                │
SampleSend-32   3.480m ± 9%   1.397m ± 6%  -59.86% (p=0.000 n=10)   1.356m ± 5%  -61.03% (p=0.000 n=10)

              │     main      │               main-v2                │                 vt20                 │
              │     B/op      │     B/op      vs base                │     B/op      vs base                │
SampleSend-32   648.03Ki ± 6%   55.75Ki ± 7%  -91.40% (p=0.000 n=10)   54.55Ki ± 7%  -91.58% (p=0.000 n=10)

              │    main    │              main-v2               │                vt20                │
              │ allocs/op  │ allocs/op   vs base                │ allocs/op   vs base                │
SampleSend-32   887.0 ± 5%   425.5 ± 6%  -52.03% (p=0.000 n=10)   585.0 ± 7%  -34.05% (p=0.000 n=10)
```