@bboreham 2024-09-29
I downloaded the code (commit 9218e8c677ef6adcffd6303f3556252aae2c4465) and got this test failure:
```
--- FAIL: TestSampleDelivery (1.51s)
    --- FAIL: TestSampleDelivery/prometheus.WriteRequest-samples,_exemplars,_and_histograms (0.50s)
panic: runtime error: index out of range [0] with length 0 [recovered]
        panic: runtime error: index out of range [0] with length 0
```

I see you mention a number of points at "Tests to fix:", but apparently not this one.