@bboreham 2025-07-08
Hello from the bug-scrub again!

> it would be more helpful to address the questions that I posted above in the description

I'm sorry, I could not detect any questions on a quick read.  I used the browser search function for "?".
Please can you clarify what you would like answered.